moved join-flow-steps to join flow dir
